feat: configure spinner inside adapter
